Removed support to .gmv, .cas, .dat and .eps files.

+++ Sources/Convert/Cgns_Mod.f90
@@ -8,7 +8,7 @@
 !==============================================================================!
 
   ! file
-  integer*8         :: file_id
+  integer           :: file_id
   character(len=80) :: file_name
   logical           :: verbose = .false.
 
@@ -16,20 +16,20 @@
   !   Element section   !
   !---------------------!
   type Cgns_Section_Type
-    character(len=80)      :: name
-    integer*8, allocatable :: cell_type
-    integer*8              :: first_cell
-    integer*8              :: last_cell
-    integer*8              :: parent_flag
+    character(len=80)    :: name
+    integer, allocatable :: cell_type
+    integer              :: first_cell
+    integer              :: last_cell
+    integer              :: parent_flag
   end type
 
   !-------------------------!
   !   Boundary conditions   ! -> it is similar to Bnd_Cond in ../Share :-(
   !-------------------------!
   type Cgns_Bnd_Cond_Type
-    character(len=80)      :: name
-    integer*8, allocatable :: color
-    integer*8, allocatable :: n_nodes
+    character(len=80)    :: name
+    integer, allocatable :: color
+    integer, allocatable :: n_nodes
   end type
 
   !------------!
@@ -37,49 +37,49 @@
   !------------!
   type Cgns_Block_Type
     character(len=80)                     :: name
-    integer*8                             :: type
-    integer*8                             :: mesh_info(3)
-    integer*8                             :: n_sects      
+    integer                               :: type
+    integer                               :: mesh_info(3)
+    integer                               :: n_sects      
     type(Cgns_Section_Type), allocatable  :: section(:)
-    integer*8                             :: n_bnd_conds
+    integer                               :: n_bnd_conds
     type(Cgns_Bnd_Cond_Type), allocatable :: bnd_cond(:)
-    integer*8                             :: n_coords
+    integer                               :: n_coords
     character(len=80)                     :: coord_name(3)
   end type
 
   !----------!
   !   Base   ! -> contains blocks
   !----------!
-  integer*8 :: n_bases
+  integer :: n_bases
   type Cgns_Base_Type
     character(len=80)                  :: name
-    integer*8                          :: cell_dim
-    integer*8                          :: phys_dim
-    integer*8                          :: n_blocks
+    integer                            :: cell_dim
+    integer                            :: phys_dim
+    integer                            :: n_blocks
     type(Cgns_Block_Type), allocatable :: block(:)
   end type
   type(Cgns_Base_Type), allocatable :: cgns_base(:)
 
   ! Some global counters (this is a bit ugly)
-  integer*8 :: cnt_nodes
-  integer*8 :: cnt_cells
-  integer*8 :: cnt_blocks     ! probably not needed
-  integer*8 :: cnt_bnd_cells
+  integer :: cnt_nodes
+  integer :: cnt_cells
+  integer :: cnt_blocks     ! probably not needed
+  integer :: cnt_bnd_cells
 
-  integer*8 :: cnt_hex
-  integer*8 :: cnt_pyr
-  integer*8 :: cnt_wed
-  integer*8 :: cnt_tet
-  integer*8 :: cnt_qua
-  integer*8 :: cnt_tri
+  integer :: cnt_hex
+  integer :: cnt_pyr
+  integer :: cnt_wed
+  integer :: cnt_tet
+  integer :: cnt_qua
+  integer :: cnt_tri
 
-  integer*8 :: cnt_x
-  integer*8 :: cnt_y
-  integer*8 :: cnt_z
+  integer :: cnt_x
+  integer :: cnt_y
+  integer :: cnt_z
 
   ! Block-wise counter of boundary cells
-  integer*8 :: cnt_block_bnd_cells  ! probably not needed  
-  integer*8 :: cnt_bnd_conds
+  integer :: cnt_block_bnd_cells  ! probably not needed  
+  integer :: cnt_bnd_conds
   character(len=80) :: bnd_cond_names(1024)
 
   contains
