+++ Sources/Convert/Cgns_Mod/Read_Number_Of_Bnd_Conds_In_Block.f90
@@ -5,12 +5,12 @@
 !------------------------------------------------------------------------------!
   implicit none
 !---------------------------------[Arguments]----------------------------------!
-  integer*8 :: base, block
+  integer :: base, block
 !-----------------------------------[Locals]-----------------------------------!
-  integer*8 :: base_id      ! base index number
-  integer*8 :: block_id     ! block index number
-  integer*8 :: n_bnd_conds  ! number of boundary conditions in block
-  integer*8 :: error
+  integer :: base_id      ! base index number
+  integer :: block_id     ! block index number
+  integer :: n_bnd_conds  ! number of boundary conditions in block
+  integer :: error
 !==============================================================================!
 
   ! Set input parameters
