+++ Sources/Convert/Cgns_Mod/Read_Section_Info.f90
@@ -5,19 +5,19 @@
 !------------------------------------------------------------------------------!
   implicit none
 !---------------------------------[Arguments]----------------------------------!
-  integer*8 :: base, block, sect
+  integer :: base, block, sect
 !-----------------------------------[Locals]-----------------------------------!
-  integer*8         :: base_id       ! base index number
-  integer*8         :: block_id      ! block index number
-  integer*8         :: sect_id       ! element section index
+  integer           :: base_id       ! base index number
+  integer           :: block_id      ! block index number
+  integer           :: sect_id       ! element section index
   character(len=80) :: sect_name     ! name of the Elements_t node
-  integer*8         :: cell_type     ! types of elements in the section
-  integer*8         :: first_cell    ! index of first element
-  integer*8         :: last_cell     ! index of last element
-  integer*8         :: n_bnd         ! index of last boundary element
-  integer*8         :: parent_flag   ! are the parent cells stored (I guess)
-  integer*8         :: error
-  integer*8         :: cnt, bc
+  integer           :: cell_type     ! types of elements in the section
+  integer           :: first_cell    ! index of first element
+  integer           :: last_cell     ! index of last element
+  integer           :: n_bnd         ! index of last boundary element
+  integer           :: parent_flag   ! are the parent cells stored (I guess)
+  integer           :: error
+  integer           :: cnt, bc
 !==============================================================================!
 
   ! Set input parameters
